chore: separate styles css file into folders
