+++ rest-hub-client/src/components/ui/Divider.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import styles from '@/styles/utils.module.css';
+import styles from '@/styles/utils/utils.module.css';
 
 export default function Divider() {
   return (
